- Moved flattenMenuItems to IColumnMenuFactory for better structure  - Updated ColumnMenuFactory to implement IColumnMenuFactory  - Removed redundant flattenMenuItems function from agColumn.ts  - Added colMenuFactory to context for improved dependency injection
