+++ packages/ag-grid-community/src/agStack/interfaces/iColumnMenuFactory.ts
@@ -1,0 +1,8 @@
+import type { DefaultMenuItem, MenuItemDef } from '../../interfaces/menuItem';
+
+export interface IColumnMenuFactory {
+    /**
+     * Returns a flat set of provided menu items names
+     */
+    flattenMenuItems(columnMainMenuItems: (DefaultMenuItem | MenuItemDef)[]): string[];
+}
