+++ packages/ag-grid-community/src/main.ts
@@ -253,6 +253,7 @@ export {
     AgComponent as _AgComponent,
     AgComponentEvent as _AgComponentEvent,
 } from './agStack/interfaces/agComponent';
+export { IColumnMenuFactory } from './agStack/interfaces/iColumnMenuFactory';
 export { AgSingletonBeanClass as _AgSingletonBeanClass, AgContext as _AgContext } from './agStack/core/agContext';
 export {
     BaseEnvironment as _BaseEnvironment,
