+++ packages/ag-grid-enterprise/src/menu/columnMenuFactory.ts
@@ -1,4 +1,11 @@
-import type { AgColumn, AgProvidedColumnGroup, DefaultMenuItem, MenuItemDef, NamedBean } from 'ag-grid-community';
+import type {
+    AgColumn,
+    AgProvidedColumnGroup,
+    DefaultMenuItem,
+    IColumnMenuFactory,
+    MenuItemDef,
+    NamedBean,
+} from 'ag-grid-community';
 import {
     BeanStub,
     _addGridCommonParams,
@@ -14,7 +21,7 @@ import { MenuList } from '../widgets/menuList';
 import type { MenuItemMapper } from './menuItemMapper';
 import { MENU_ITEM_SEPARATOR, _removeRepeatsFromArray } from './menuItemMapper';
 
-export class ColumnMenuFactory extends BeanStub implements NamedBean {
+export class ColumnMenuFactory extends BeanStub implements NamedBean, IColumnMenuFactory {
     beanName = 'colMenuFactory' as const;
 
     public createMenu(
@@ -80,6 +87,29 @@ export class ColumnMenuFactory extends BeanStub implements NamedBean {
         _removeRepeatsFromArray(result, MENU_ITEM_SEPARATOR);
 
         return result;
+    }
+
+    /**
+     * Returns a flat set of provided menu items names
+     */
+    public flattenMenuItems(columnMainMenuItems: (DefaultMenuItem | MenuItemDef)[]): string[] {
+        const mapper = (arr: string[], item: DefaultMenuItem | MenuItemDef) => {
+            if (typeof item === 'string') {
+                arr.push(item);
+            }
+            if (typeof item === 'object') {
+                if (item.subMenu) {
+                    // eslint-disable-next-line sonarjs/no-ignored-return
+                    item.subMenu.reduce(mapper, arr);
+                } else {
+                    arr.push(item.name);
+                }
+            }
+
+            return arr;
+        };
+
+        return columnMainMenuItems.reduce(mapper, []);
     }
 
     private getDefaultMenuOptions(column: AgColumn | null): DefaultMenuItem[] {
